+++ src/main/java/com/idark/valoria/registries/item/types/ranged/GunpowderCharge.java
@@ -1,15 +1,12 @@
 package com.idark.valoria.registries.item.types.ranged;
 
-import net.minecraft.ChatFormatting;
-import net.minecraft.network.chat.CommonComponents;
-import net.minecraft.network.chat.Component;
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.level.Level;
-import org.jetbrains.annotations.NotNull;
+import net.minecraft.*;
+import net.minecraft.network.chat.*;
+import net.minecraft.world.item.*;
+import net.minecraft.world.level.*;
+import org.jetbrains.annotations.*;
 
-import java.util.List;
+import java.util.*;
 
 import static net.minecraft.world.item.ItemStack.ATTRIBUTE_MODIFIER_FORMAT;
 
@@ -35,6 +32,8 @@ public class GunpowderCharge extends Item{
     @Override
     public void appendHoverText(@NotNull ItemStack stack, Level world, @NotNull List<Component> tooltip, @NotNull TooltipFlag flags){
         super.appendHoverText(stack, world, tooltip, flags);
+        tooltip.add(Component.translatable("tooltip.valoria.gunpowder_charge").withStyle(ChatFormatting.GRAY));
+        tooltip.add(Component.empty());
         tooltip.add(CommonComponents.space().append(Component.translatable("attribute.modifier.equals.0", ATTRIBUTE_MODIFIER_FORMAT.format(damage), Component.translatable("attribute.name.generic.attack_damage")).withStyle(ChatFormatting.DARK_GREEN)));
         tooltip.add(CommonComponents.space().append(Component.translatable("attribute.modifier.equals.0", ATTRIBUTE_MODIFIER_FORMAT.format(knockback), Component.translatable("attribute.name.generic.attack_knockback")).withStyle(ChatFormatting.DARK_GREEN)));
         tooltip.add(CommonComponents.space().append(Component.translatable("attribute.modifier.equals.0", ATTRIBUTE_MODIFIER_FORMAT.format(radius), Component.translatable("attribute.valoria.attack_radius")).withStyle(ChatFormatting.DARK_GREEN)));